fix : 선거 비공개 상태에서만 후보자 삭제가 가능하도록 변경

+++ src/main/java/com/keeper/homepage/domain/election/application/AdminElectionService.java
@@ -94,7 +94,10 @@ public class AdminElectionService {
   @Transactional
   public void deleteCandidate(long electionId, long candidateId) {
     ElectionCandidate electionCandidate = validCandidateService.findById(candidateId);
-    validElectionFindService.findById(electionId);
+    Election election = validElectionFindService.findById(electionId);
+    if (election.isAvailable()) {
+      throw new BusinessException(electionId, "electionId", ErrorCode.ELECTION_CANDIDATE_CANNOT_DELETE);
+    }
 
     electionCandidateRepository.delete(electionCandidate);
   }
